requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
distro==1.9.0
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
googlesearch-python==1.2.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.14.1
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.32.2
tenacity==8.2.3
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
